refactor: Move functions used for test only to test code
