+++ src/main/kotlin/lotto/util/Strings.kt
@@ -1,7 +1,8 @@
 package lotto.util
 
-fun String.toBlankRemovedIntList() =
-    this.split(",").map { it.trim().toInt() }
-
-fun String.toBlankRemovedIntSet() =
-    this.toBlankRemovedIntList().toSet()
+fun String.isInt() = try {
+    this.toInt()
+    true
+} catch (e: NumberFormatException) {
+    false
+}
