+++ src/test/kotlin/lotto/model/RangeLottoBuilderTest.kt
@@ -3,7 +3,6 @@ package lotto.model
 import lotto.model.LottoEvaluator.countOfMatchNumber
 import lotto.model.data.Lotto.Companion.toLotto
 import lotto.model.data.LottoNumber
-import lotto.model.data.LottoNumber.Companion.toLottoNumber
 import lotto.model.data.Lottos
 import lotto.model.data.Policy645
 import lotto.model.data.Result
@@ -11,6 +10,7 @@ import lotto.model.data.Results
 import lotto.model.data.Statistics
 import lotto.model.data.Winning
 import lotto.model.data.WinningLotto.Companion.toWinningLotto
+import lotto.model.data.toLottoNumber
 import lotto.view.input.LottosInputView
 import lotto.view.input.ManualLottosInputView
 import lotto.view.output.ConsoleOutputView
